Re-usable TableStorageQuery + JSON msgs to UI
https://github.com/Pure-Battlefield/gswat/issues/45 - ### Combined some
of the code that was being duplicated to query the ServerSettings from
the Azure Storage, and gave it dynamic partitionKey and rowKey values in
light of the upcoming change in the StorageSchema.

- Changed the Core.Connect function to return a string either containing
the Exception message that was caught, or a custom "Connected to."
string (we need to implement a better way of verifying that we are
indeed connected in the CommLayer) and changed the
ValuesController.SetServerInfo method to return these as a JSON
serialized object to the front-end.

- Added a GetServerSettings() method in the ValuesController that uses
the above-mentioned StorageQuery to get the ServerSettings (only IP and
Port right now), the idea being that we can call it for each server the
user is looking at. This is also served as a JSON object to the
Front-End to let Enzo do with it what he pleases (HTTPGet). I've added
an <Auth> flag in the comment, because this action should probably
require authorization once we implement it.

Because of the current Front-End not being connected to the actions, use
direct URLs (e.g. /api/values/getserversettings) to test it.